refactor: finishing up
